refactor: move datasource logic to provider/publiccloud
