--- internal/provider/data_sources/public_cloud/instances/data_source.go
@@ -1,19 +1,0 @@
-package instances
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/client"
-)
-
-var (
-	_ datasource.DataSource              = &instancesDataSource{}
-	_ datasource.DataSourceWithConfigure = &instancesDataSource{}
-)
-
-func NewInstancesDataSource() datasource.DataSource {
-	return &instancesDataSource{}
-}
-
-type instancesDataSource struct {
-	client client.Client
-}
