--- internal/provider/data_sources/public_cloud/instances/doc.go
@@ -1,2 +1,0 @@
-// Package instances implements functionality for public_cloud data sources.
-package instances
