--- internal/provider/data_sources/public_cloud/instances/metadata.go
@@ -1,15 +1,0 @@
-package instances
-
-import (
-	"context"
-
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-)
-
-func (d *instancesDataSource) Metadata(
-	_ context.Context,
-	req datasource.MetadataRequest,
-	resp *datasource.MetadataResponse,
-) {
-	resp.TypeName = req.ProviderTypeName + "_public_cloud_instances"
-}
