--- internal/provider/data_sources/public_cloud/model/doc.go
@@ -1,2 +1,0 @@
-// Package model implements data models for public_cloud data sources.
-package model
