+++ internal/provider/provider.go
@@ -12,7 +12,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/client"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/data_sources/public_cloud/instances"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/resources/public_cloud/instance"
 )
 
@@ -157,7 +157,7 @@ func (p *leasewebProvider) Configure(
 
 func (p *leasewebProvider) DataSources(_ context.Context) []func() datasource.DataSource {
 	return []func() datasource.DataSource{
-		instances.NewInstancesDataSource,
+		publiccloud.NewInstancesDataSource,
 		NewDedicatedServerDataSource,
 		NewDedicatedServersDataSource,
 		NewDedicatedServerControlPanelsDataSource,
