+++ internal/provider/publiccloud/data_source_instances_test.go
@@ -1,4 +1,4 @@
-package instances
+package publiccloud
 
 import (
 	"context"
