+++ internal/provider/publiccloud/doc.go
@@ -1,0 +1,2 @@
+// Package publiccloud implements functionality for public_cloud data sources & resources.
+package publiccloud
