+++ internal/provider/publiccloud/models/datasource/doc.go
@@ -1,0 +1,2 @@
+// Package datasource stores the publiccloud data sources.
+package datasource
