+++ internal/provider/publiccloud/models/datasource/instances.go
@@ -1,4 +1,4 @@
-package model
+package datasource
 
 type Instances struct {
 	Instances []Instance `tfsdk:"instances"`
