+++ internal/provider/publiccloud/service/public_cloud/data_adapters/to_data_source_model/adapter.go
@@ -3,12 +3,12 @@ package to_data_source_model
 import (
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 	"github.com/leaseweb/leaseweb-go-sdk/publicCloud"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/data_sources/public_cloud/model"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/models/datasource"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/service/public_cloud/data_adapters/shared"
 )
 
-func AdaptInstances(sdkInstances []publicCloud.Instance) model.Instances {
-	var instances model.Instances
+func AdaptInstances(sdkInstances []publicCloud.Instance) datasource.Instances {
+	var instances datasource.Instances
 
 	for _, sdkInstance := range sdkInstances {
 		instance := adaptInstance(sdkInstance)
@@ -18,8 +18,8 @@ func AdaptInstances(sdkInstances []publicCloud.Instance) model.Instances {
 	return instances
 }
 
-func adaptInstance(sdkInstance publicCloud.Instance) model.Instance {
-	return model.Instance{
+func adaptInstance(sdkInstance publicCloud.Instance) datasource.Instance {
+	return datasource.Instance{
 		Id:                  basetypes.NewStringValue(sdkInstance.Id),
 		Region:              basetypes.NewStringValue(string(sdkInstance.Region)),
 		Reference:           shared.AdaptNullableStringToStringValue(sdkInstance.Reference.Get()),
@@ -34,14 +34,14 @@ func adaptInstance(sdkInstance publicCloud.Instance) model.Instance {
 	}
 }
 
-func adaptImage(sdkImage publicCloud.Image) model.Image {
-	return model.Image{
+func adaptImage(sdkImage publicCloud.Image) datasource.Image {
+	return datasource.Image{
 		Id: basetypes.NewStringValue(sdkImage.Id),
 	}
 }
 
-func adaptIps(sdkIps []publicCloud.Ip) []model.Ip {
-	var ips []model.Ip
+func adaptIps(sdkIps []publicCloud.Ip) []datasource.Ip {
+	var ips []datasource.Ip
 	for _, sdkIp := range sdkIps {
 		ips = append(
 			ips,
@@ -52,14 +52,14 @@ func adaptIps(sdkIps []publicCloud.Ip) []model.Ip {
 	return ips
 }
 
-func adaptIp(sdkIp publicCloud.Ip) model.Ip {
-	return model.Ip{
+func adaptIp(sdkIp publicCloud.Ip) datasource.Ip {
+	return datasource.Ip{
 		Ip: basetypes.NewStringValue(sdkIp.Ip),
 	}
 }
 
-func adaptContract(sdkContract publicCloud.Contract) model.Contract {
-	return model.Contract{
+func adaptContract(sdkContract publicCloud.Contract) datasource.Contract {
+	return datasource.Contract{
 		BillingFrequency: basetypes.NewInt64Value(int64(sdkContract.BillingFrequency)),
 		Term:             basetypes.NewInt64Value(int64(sdkContract.Term)),
 		Type:             basetypes.NewStringValue(string(sdkContract.Type)),
